Moved battery optimization checks to main activity and added dialog
