+++ app/app/src/main/java/com/octo4a/ui/MainActivity.kt
@@ -1,21 +1,37 @@
 package com.octo4a.ui
 
+import android.app.AlertDialog
+import android.app.UiModeManager
+import android.content.ActivityNotFoundException
+import android.content.DialogInterface
 import android.content.Intent
+import android.content.res.Configuration
+import android.net.Uri
+import android.os.Build
 import android.os.Bundle
+import android.os.PowerManager
+import android.provider.Settings
 import android.view.Menu
 import android.view.MenuItem
+import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.LifecycleService
 import androidx.navigation.findNavController
 import androidx.navigation.ui.setupWithNavController
 import com.octo4a.R
+import com.octo4a.repository.LoggerRepository
 import com.octo4a.serial.VirtualSerialDriver
 import com.octo4a.service.OctoPrintService
 import com.octo4a.ui.fragments.TerminalSheetDialog
+import com.octo4a.utils.preferences.MainPreferences
 import kotlinx.android.synthetic.main.activity_main.*
 import org.koin.android.ext.android.inject
 
 class MainActivity : AppCompatActivity() {
+    private val logger: LoggerRepository by inject()
     private val vsp: VirtualSerialDriver by inject()
+    private val mainPreferences: MainPreferences by inject()
+    private val pm  by lazy { getSystemService(LifecycleService.POWER_SERVICE) as PowerManager }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -23,6 +39,48 @@ class MainActivity : AppCompatActivity() {
         setSupportActionBar(my_toolbar)
         bottomNavigationBar.setupWithNavController(findNavController(R.id.navHost))
         vsp.updateDevicesList(OctoPrintService.BROADCAST_SERVICE_USB_GOT_ACCESS)
+
+        // Required for acquiring wakelock
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M && !pm.isIgnoringBatteryOptimizations(packageName) && mainPreferences.warnDisableBatteryOptimization) {
+            showBatteryOptimizationDialog()
+        }
+    }
+
+    private fun showBatteryOptimizationDialog() {
+        val uiModeManager: UiModeManager = getSystemService(UI_MODE_SERVICE) as UiModeManager
+
+        val alertDialog: AlertDialog? = this.let {
+            val builder = AlertDialog.Builder(it)
+            builder.apply {
+                setTitle(R.string.disable_battery_optimization)
+                setMessage(R.string.disable_battery_optimization_msg)
+                setNegativeButton(R.string.action_never_ask_again) { dialog, id ->
+                    mainPreferences.warnDisableBatteryOptimization = false
+                }
+                setNeutralButton(R.string.action_later) { dialog, id ->
+                    // User cancelled the dialog
+                }
+
+                if (uiModeManager.currentModeType == Configuration.UI_MODE_TYPE_NORMAL) {
+                    setPositiveButton(R.string.action_open_settings) { dialog, id ->
+                        val whitelist = Intent()
+                        whitelist.action = Settings.ACTION_REQUEST_IGNORE_BATTERY_OPTIMIZATIONS
+                        whitelist.data = Uri.parse("package:$packageName")
+                        whitelist.flags = Intent.FLAG_ACTIVITY_NEW_TASK
+                        try {
+                            startActivity(whitelist)
+                        } catch (e: ActivityNotFoundException) {
+                            logger.log(this) { "failed to open battery optimization dialog" }
+                            Toast.makeText(this@MainActivity, "Failed to open battery optimization settings", Toast.LENGTH_LONG).show()
+                        }
+                    }
+                }
+            }
+
+            // Create the AlertDialog
+            builder.create()
+        }
+        alertDialog?.show()
     }
 
     // @TODO: refactor to non deprecated api
